+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Project Field Row
 - Notifications in project dashboard
 - Fetch project states
+- DataSHIELD status